feat(RomPageLauncherFactoryWrapper):添加包装类简化适配方法实现
