+++ agent/src/main/java/cn/junhua/android/permission/rom/qihu360/Qihu360PageLauncherFactory.java
@@ -3,7 +3,7 @@ package cn.junhua.android.permission.rom.qihu360;
 import android.os.Build;
 
 import cn.junhua.android.permission.rom.PageLauncher;
-import cn.junhua.android.permission.rom.RomPageLauncherFactory;
+import cn.junhua.android.permission.rom.RomPageLauncherFactoryWrapper;
 import cn.junhua.android.permission.utils.RomUtils;
 
 /**
@@ -12,21 +12,11 @@ import cn.junhua.android.permission.utils.RomUtils;
  * @author junhua.lin@jinfuzi.com<br/>
  * CREATED 2019/6/19 10:09
  */
-public class Qihu360PageLauncherFactory implements RomPageLauncherFactory {
+public class Qihu360PageLauncherFactory extends RomPageLauncherFactoryWrapper {
     @Override
     public boolean check() {
         return RomUtils.checkManufacturer("qiku")
                 || RomUtils.checkManufacturer("360");
-    }
-
-    @Override
-    public PageLauncher createInstallLauncher() {
-        return null;
-    }
-
-    @Override
-    public PageLauncher createNotifyLauncher() {
-        return null;
     }
 
     @Override
@@ -38,13 +28,4 @@ public class Qihu360PageLauncherFactory implements RomPageLauncherFactory {
         return null;
     }
 
-    @Override
-    public PageLauncher createWriteSettingsLauncher() {
-        return null;
-    }
-
-    @Override
-    public PageLauncher createAppDetailLauncher() {
-        return null;
-    }
 }
