+++ agent/src/main/java/cn/junhua/android/permission/rom/xiaomi/XiaomiPageLauncherFactory.java
@@ -3,7 +3,7 @@ package cn.junhua.android.permission.rom.xiaomi;
 import android.os.Build;
 
 import cn.junhua.android.permission.rom.PageLauncher;
-import cn.junhua.android.permission.rom.RomPageLauncherFactory;
+import cn.junhua.android.permission.rom.RomPageLauncherFactoryWrapper;
 import cn.junhua.android.permission.utils.RomUtils;
 
 /**
@@ -12,7 +12,7 @@ import cn.junhua.android.permission.utils.RomUtils;
  * @author junhua.lin@jinfuzi.com<br/>
  * CREATED 2019/6/19 10:09
  */
-public class XiaomiPageLauncherFactory implements RomPageLauncherFactory {
+public class XiaomiPageLauncherFactory extends RomPageLauncherFactoryWrapper {
     @Override
     public boolean check() {
         return RomUtils.checkManufacturer("xiaomi")
@@ -22,26 +22,11 @@ public class XiaomiPageLauncherFactory implements RomPageLauncherFactory {
     }
 
     @Override
-    public PageLauncher createInstallLauncher() {
-        return null;
-    }
-
-    @Override
-    public PageLauncher createNotifyLauncher() {
-        return null;
-    }
-
-    @Override
     public PageLauncher createOverlayLauncher() {
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M
                 && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
             return new KXiaomiOverlayPageLauncher();
         }
-        return null;
-    }
-
-    @Override
-    public PageLauncher createWriteSettingsLauncher() {
         return null;
     }
